+++ src/main/java/in/koreatech/koin/domain/student/service/StudentService.java
@@ -51,7 +51,7 @@ import in.koreatech.koin.domain.user.repository.UserTokenRedisRepository;
 import in.koreatech.koin.domain.user.service.RefreshTokenService;
 import in.koreatech.koin.domain.user.service.UserService;
 import in.koreatech.koin.domain.user.service.UserValidationService;
-import in.koreatech.koin.domain.user.service.UserVerificationService;
+import in.koreatech.koin.domain.user.verification.service.UserVerificationService;
 import in.koreatech.koin.domain.student.model.StudentPasswordChangeData;
 import in.koreatech.koin.domain.student.model.StudentRegistrationData;
 import in.koreatech.koin.infrastructure.email.service.MailService;
