+++ src/main/java/in/koreatech/koin/domain/user/controller/UserApi.java
@@ -47,7 +47,7 @@ import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 
-@Tag(name = "(Normal) User: 회원", description = "회원 관련 API")
+@Tag(name = "(Normal) User: 유저", description = "유저 관련 API")
 public interface UserApi {
 
     @ApiResponses(
@@ -188,93 +188,6 @@ public interface UserApi {
 
     @ApiResponses(
         value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "400", description = "이메일 양식 오류", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "409", description = "이메일 중복", content = @Content(schema = @Schema(hidden = true)))
-        }
-    )
-    @Operation(
-        summary = "이메일 중복 체크",
-        description = "입력한 이메일이 중복인지 확인합니다."
-    )
-    @GetMapping("/user/check/email")
-    ResponseEntity<Void> checkUserEmailExist(
-        @ParameterObject @ModelAttribute("address")
-        @Valid CheckEmailDuplicationRequest request
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "400", description = "전화번호 양식 오류", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "409", description = "전화번호 중복", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(
-        summary = "전화번호 중복 체크",
-        description = "입력한 전화번호가 중복인지 확인합니다."
-    )
-    @GetMapping("/user/check/phone")
-    ResponseEntity<Void> checkPhoneNumberExist(
-        @ParameterObject @ModelAttribute("phone")
-        @Valid CheckPhoneDuplicationRequest request
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "400", description = "아이디 양식 오류", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "409", description = "아이디 중복", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(
-        summary = "아이디 중복 체크",
-        description = "입력한 아이디가 중복인지 확인합니다."
-    )
-    @GetMapping("/user/check/id")
-    ResponseEntity<Void> checkDuplicatedLoginId(
-        @ParameterObject @ModelAttribute("id")
-        @Valid CheckLoginIdDuplicationRequest request
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "400", description = "닉네임 양식 오류", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "409", description = "닉네임 중복", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(
-        summary = "닉네임 중복 체크",
-        description = "입력한 닉네임이 중복인지 확인합니다."
-    )
-    @GetMapping("/user/check/nickname")
-    ResponseEntity<Void> checkDuplicationOfNickname(
-        @ParameterObject @ModelAttribute("nickname")
-        @Valid CheckNicknameDuplicationRequest request
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "400", description = "적절한 비밀번호가 아님", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "401", description = "로그인 하지 않음", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", description = "권한 없음", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(
-        summary = "비밀번호 검증",
-        description = "적절한 비밀번호인지 검증합니다."
-    )
-    @SecurityRequirement(name = "Jwt Authentication")
-    @PostMapping("/user/check/password")
-    ResponseEntity<Void> checkPassword(
-        @Valid @RequestBody CheckUserPasswordRequest request,
-        @Auth(permit = {GENERAL, STUDENT, OWNER, COOP, COUNCIL}) Integer userId
-    );
-
-    @ApiResponses(
-        value = {
             @ApiResponse(responseCode = "200", description = "로그인 확인"),
             @ApiResponse(responseCode = "401", description = "로그인 하지 않음", content = @Content(schema = @Schema(hidden = true))),
         }
@@ -289,63 +202,6 @@ public interface UserApi {
         @ParameterObject @ModelAttribute(value = "access_token")
         @Valid UserAccessTokenRequest request
     );
-
-    @ApiResponses({
-        @ApiResponse(responseCode = "200", description = "사용자 ID 존재"),
-        @ApiResponse(responseCode = "404", description = "사용자 ID 없음", content = @Content(schema = @Schema(hidden = true)))
-    })
-    @Operation(
-        summary = "로그인 ID 존재 여부 확인",
-        description = "입력한 로그인 ID가 존재하는지 확인합니다."
-    )
-    @PostMapping("/user/id/exists")
-    ResponseEntity<Void> existsByUserId(@Valid @RequestBody ExistsByUserIdRequest request);
-
-    @ApiResponses({
-        @ApiResponse(responseCode = "200", description = "전화번호 존재"),
-        @ApiResponse(responseCode = "404", description = "전화번호 없음", content = @Content(schema = @Schema(hidden = true)))
-    })
-    @Operation(
-        summary = "전화번호 존재 여부 확인",
-        description = "입력한 전화번호로 가입된 계정이 존재하는지 확인합니다."
-    )
-    @PostMapping("/user/phone/exists")
-    ResponseEntity<Void> existsByPhoneNumber(@Valid @RequestBody ExistsByPhoneRequest request);
-
-    @ApiResponses({
-        @ApiResponse(responseCode = "200", description = "이메일 존재"),
-        @ApiResponse(responseCode = "404", description = "이메일 없음", content = @Content(schema = @Schema(hidden = true)))
-    })
-    @Operation(
-        summary = "이메일 존재 여부 확인",
-        description = "입력한 이메일 주소로 가입된 계정이 존재하는지 확인합니다."
-    )
-    @PostMapping("/user/email/exists")
-    ResponseEntity<Void> existsByEmail(@Valid @RequestBody ExistsByEmailRequest request);
-
-    @ApiResponses({
-        @ApiResponse(responseCode = "200", description = "ID와 전화번호 일치"),
-        @ApiResponse(responseCode = "400", description = "ID와 전화번호 불일치", content = @Content(schema = @Schema(hidden = true))),
-        @ApiResponse(responseCode = "404", description = "ID 없음", content = @Content(schema = @Schema(hidden = true)))
-    })
-    @Operation(
-        summary = "로그인 ID와 전화번호 일치 여부 확인",
-        description = "입력한 로그인 ID와 전화번호가 일치하는지 확인합니다."
-    )
-    @PostMapping("/users/id/match/phone")
-    ResponseEntity<Void> matchUserIdWithPhoneNumber(@Valid @RequestBody MatchUserIdWithPhoneNumberRequest request);
-
-    @ApiResponses({
-        @ApiResponse(responseCode = "200", description = "ID와 이메일 일치"),
-        @ApiResponse(responseCode = "400", description = "ID와 이메일 불일치", content = @Content(schema = @Schema(hidden = true))),
-        @ApiResponse(responseCode = "404", description = "ID 없음", content = @Content(schema = @Schema(hidden = true)))
-    })
-    @Operation(
-        summary = "로그인 ID와 이메일 일치 여부 확인",
-        description = "입력한 로그인 ID와 이메일 주소가 일치하는지 확인합니다."
-    )
-    @PostMapping("/users/id/match/email")
-    ResponseEntity<Void> matchUserIdWithEmail(@Valid @RequestBody MatchUserIdWithEmailRequest request);
 
     @ApiResponses({
         @ApiResponse(responseCode = "200", description = "ID 조회 성공"),
