+++ src/test/java/in/koreatech/koin/acceptance/UserApiTest.java
@@ -27,7 +27,7 @@ import in.koreatech.koin.domain.user.model.User;
 import in.koreatech.koin.domain.user.model.UserGender;
 import in.koreatech.koin.domain.user.model.UserVerificationStatus;
 import in.koreatech.koin.domain.user.repository.UserRepository;
-import in.koreatech.koin.domain.user.repository.UserVerificationStatusRedisRepository;
+import in.koreatech.koin.domain.user.verification.repository.UserVerificationStatusRedisRepository;
 import in.koreatech.koin.fixture.DepartmentFixture;
 import in.koreatech.koin.fixture.UserFixture;
 import in.koreatech.koin.infrastructure.naver.service.NaverSmsService;
